requirements.txt
discord.py
python-dotenv
pymongo
deep-translator